requirements.txt
pillow
opencv-contrib-python==4.0.0.21
web.py==0.40.dev0
torch torchvision